Fixed a bug in the calculation of the spin axis.  We assume the spin axis has mostly phi and no theta (in terms of spherical coordinates).  However, the spin axis simply isn't phi because theta can be -pi to pi.  When theta is 0 to pi, the spin axis should be positive.  Conversely, when theta is -pi to 0, the spin axis should be negative.  Keep in mind, phi is ALWAYS positive in spherical coordiantes (at least in the calculation as phi = acos(z / r)).
Note for future, there is still a bug/odd behavior here.  When the program gets a scenario with a theta near 0 or pi AND a large phi, phi no longer makes sense.  e.g. a theta of 0 and a phi of 45 degrees returns a spin axis of 45 degrees.  I don't think this is want we want.

Also, a bug in the test program, main.c, where I was passing a line of 89/2 to a double.  This wasn't casting correctly, need 89.0/2.0 to force a double divide instead of an integer divide following be a conversion to double.